requirements.txt
dnspython==2.8.0
googleapis-common-protos==1.70.0
grpclib==0.4.8
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
multidict==6.6.4
numpy==2.2.6
opencv-python==4.12.0.88
protobuf==5.29.5
pymongo==4.15.1
typing_extensions==4.15.0
viam-sdk==0.56.0

typing-extensions
